+++ src/utils/api.ts
@@ -239,10 +239,10 @@ function invalidCsrfResponse(res) {
  */
 export async function validateAddress(req, res, appObj = app) {
   const tokenObject = appObj["tokenObject"];
-  // const { csrfTokenValid } = utils.getCsrfToken(req, res);
-  // if (!csrfTokenValid) {
-  //   return invalidCsrfResponse(res);
-  // }
+  const { csrfTokenValid } = utils.getCsrfToken(req, res);
+  if (!csrfTokenValid) {
+    return invalidCsrfResponse(res);
+  }
   if (tokenObject && tokenObject?.access_token) {
     const token = tokenObject.access_token;
     const addressRequest: AddressAPIRequestData = {
@@ -294,10 +294,10 @@ export async function validateUsername(
   appObj = app
 ) {
   const tokenObject = appObj["tokenObject"];
-  // const { csrfTokenValid } = utils.getCsrfToken(req, res);
-  // if (!csrfTokenValid) {
-  //   return invalidCsrfResponse(res);
-  // }
+  const { csrfTokenValid } = utils.getCsrfToken(req, res);
+  if (!csrfTokenValid) {
+    return invalidCsrfResponse(res);
+  }
   if (tokenObject && tokenObject?.access_token) {
     const token = tokenObject.access_token;
     const username = req.body.username;
@@ -428,10 +428,10 @@ export async function createPatron(
   appObj = app
 ) {
   const data = req.body;
-  // const { csrfTokenValid } = utils.getCsrfToken(req, res);
-  // if (!csrfTokenValid) {
-  //   return invalidCsrfResponse(res);
-  // }
+  const { csrfTokenValid } = utils.getCsrfToken(req, res);
+  if (!csrfTokenValid) {
+    return invalidCsrfResponse(res);
+  }
 
   try {
     const response = await callPatronAPI(data, createPatronUrl, appObj);